# ingester: parallelize label values cardinality computation