<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

Adapts `labelValuesCardinality` to iterate postings in parallel to obtain label values total series count.

~~Additionally, an extra benchmark has been included to compare performance with the old implementation.~~  To do performance comparison the [recently refactored](https://github.com/grafana/mimir/pull/3080) `BenchmarkIngester_LabelValuesCardinality` has been used. 

The new implementation is about `2x-3x` faster in my testing.

Run the following command to verify the results:

```
go test -run=X -bench=BenchmarkIngester_LabelValuesCardinality -benchmem -count=5 -cpu=4 ./pkg/ingester/
```

Comparison:

```
name                                                                                             old time/op    new time/op    delta
Ingester_LabelValuesCardinality/no_matchers,___name___label_with_1_value_all_series-4              8.06ms ± 4%    7.95ms ± 0%      ~     (p=1.000 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_10_label,_1M_series_each-4                        7.95ms ± 0%    2.54ms ± 1%   -68.07%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_77_label,_130K_series_each-4                      8.03ms ± 0%    2.25ms ± 1%   -71.97%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_4199,_2400_series_each-4                          11.8ms ± 0%     5.3ms ± 1%   -54.69%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_10_label,_1M_series_each-4                    274ms ± 0%      77ms ± 1%   -71.85%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_77_label,_130K_series_each-4                  831ms ± 0%     197ms ± 1%   -76.26%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_4199_label,_2400_series_each-4                835ms ± 3%     200ms ± 7%   -76.00%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___and_mod_10_matchers,_mod_4199_label,_240_series_each-4    49.8ms ± 0%    50.7ms ± 1%    +1.96%  (p=0.008 n=5+5)

name                                                                                             old alloc/op   new alloc/op   delta
Ingester_LabelValuesCardinality/no_matchers,___name___label_with_1_value_all_series-4              2.30kB ± 0%    4.80kB ± 0%  +108.46%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_10_label,_1M_series_each-4                        5.70kB ± 0%    8.63kB ± 0%   +51.41%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_77_label,_130K_series_each-4                      40.8kB ± 0%    47.0kB ± 0%   +15.31%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_4199,_2400_series_each-4                          2.27MB ± 0%    2.48MB ± 0%    +9.12%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_10_label,_1M_series_each-4                   8.64kB ± 2%   11.53kB ± 0%   +33.51%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_77_label,_130K_series_each-4                 62.1kB ± 2%    68.9kB ± 1%   +10.94%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_4199_label,_2400_series_each-4               61.7kB ± 2%    68.2kB ± 1%   +10.59%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___and_mod_10_matchers,_mod_4199_label,_240_series_each-4    3.21kB ± 0%    8.20kB ± 0%  +155.48%  (p=0.008 n=5+5)

name                                                                                             old allocs/op  new allocs/op  delta
Ingester_LabelValuesCardinality/no_matchers,___name___label_with_1_value_all_series-4                43.0 ± 0%      64.0 ± 0%   +48.84%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_10_label,_1M_series_each-4                           127 ± 0%       157 ± 0%   +23.62%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_77_label,_130K_series_each-4                         811 ± 0%       907 ± 0%   +11.94%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_4199,_2400_series_each-4                           42.3k ± 0%     46.5k ± 0%    +9.97%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_10_label,_1M_series_each-4                      203 ± 0%       232 ± 0%   +14.29%  (p=0.016 n=4+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_77_label,_130K_series_each-4                  1.30k ± 1%     1.39k ± 1%    +7.24%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_4199_label,_2400_series_each-4                1.30k ± 0%     1.39k ± 0%    +7.27%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___and_mod_10_matchers,_mod_4199_label,_240_series_each-4      75.0 ± 0%     116.0 ± 0%   +54.67%  (p=0.008 n=5+5)
```

#### Which issue(s) this PR fixes or relates to

Fixes N/A

#### Checklist

- [X] Tests updated
- [ ] Documentation added
- [X] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`